moved common/models/actions to types/actions
