+++ apps/studio/electron/preload/webview/events/publish.ts
@@ -2,7 +2,7 @@ import { ipcRenderer } from 'electron';
 import { buildLayerTree } from '../dom';
 import { getDomElement } from '../elements/helpers';
 import { WebviewChannels } from '@onlook/types/constants';
-import type { ActionElementLocation } from '/common/models/actions';
+import type { ActionElementLocation } from '@onlook/types/actions';
 import type { DomElement } from '/common/models/element';
 
 export function publishInsertElement(
