+++ apps/studio/src/lib/editor/engine/code/insert.ts
@@ -1,7 +1,7 @@
 import { twMerge } from 'tailwind-merge';
 import { getCssClasses } from './helpers';
-import type { ActionElement, ActionElementLocation } from '/common/models/actions';
-import { CodeActionType, type CodeInsert } from '/common/models/actions/code';
+import type { ActionElement, ActionElementLocation } from '@onlook/types/actions';
+import { CodeActionType, type CodeInsert } from '@onlook/types/actions/code';
 
 export function getInsertedElement(
     actionElement: ActionElement,
