+++ apps/studio/src/lib/editor/engine/element/index.ts
@@ -2,7 +2,7 @@ import { debounce } from 'lodash';
 import { makeAutoObservable } from 'mobx';
 import type { EditorEngine } from '..';
 import { escapeSelector } from '/common/helpers';
-import type { RemoveElementAction } from '/common/models/actions';
+import type { RemoveElementAction } from '@onlook/types/actions';
 import type { DomElement, WebViewElement } from '/common/models/element';
 
 export class ElementManager {
