+++ apps/studio/src/lib/editor/engine/history/index.ts
@@ -2,7 +2,7 @@ import { sendAnalytics } from '@/lib/utils';
 import { makeAutoObservable } from 'mobx';
 import type { EditorEngine } from '..';
 import { assertNever } from '/common/helpers';
-import type { Action, Change, MoveActionLocation } from '/common/models/actions';
+import type { Action, Change, MoveActionLocation } from '@onlook/types/actions';
 
 function reverse<T>(change: Change<T>): Change<T> {
     return { updated: change.original, original: change.updated };
