+++ apps/studio/src/lib/editor/engine/insert/index.ts
@@ -10,7 +10,7 @@ import type {
     ActionElementLocation,
     ActionTarget,
     InsertElementAction,
-} from '/common/models/actions';
+} from '@onlook/types/actions';
 import type { ElementPosition } from '/common/models/element';
 
 export class InsertManager {
